expand validationFunction
Hello everyone!
After working through this chapter of the documentation I propose to expand the code-snippet of the validation function for the SandwichOrder's FormBuilder.
In the style of the other examples, there should be more surrounding code to make it easier for inexperienced readers to understand where the validation function should be added to. During my first read-through, I had to pull up and search the corresponding sample project to understand where this code should live, which was different than other parts of the documentation previously.