Update Extensions.xml
fix the default value of the `token` parameter in C# to what's in the source: default(CancellationToken). CancellationToken is a struct, and can't be null. also removed the "= null" from the VB.NET signature - there is no `null` in VB (there is `Nothing`), and I don't know how to do `default()` operator on VB.NET.

Also, the CancellationToken in the published document links back to MSDN, although it is documented on MSDocs: https://docs.microsoft.com/en-us/dotnet/api/system.threading.cancellationtoken.