Fix git push error for protected CLA branch (#1080)
* Adding support statement for regional LUIS apps (#224)

Customers requesting official documentation of lack of support for non US LUIS apps for speech priming feature for EU due to GDPR requirements to keep data in EU and no workaround currently available.

* Update tutorial doc (#240)

inserted line 74, as missing it in the global.asax.cs file causes a 404 error for customers.

* Update bot-builder-dotnet-sdk-quickstart.md

* Update bot-builder-basics.md (#314)

Incorporate feedback from the grammar checking bot while making the sentence more precise at the same time.

* Update bot-builder-basics.md (#317)

Incorporate grammar bot feedback.

* Update bot-builder-concept-middleware.md (#318)

Typo. Replace "alos" with "also"

* Update bot-builder-howto-v4-state.md

1. Fixed code to use variables than hardcoded string in Startup.cs, variables were defined but not used previously.
2. Fixed typo Modift -> Modify

* Update bot-service-quickstart.md

updated to indicate that placing resources within the same region should be preferred

* Update node-send-card-buttons.js

Add a line of code to actually send the message as well.

* Misspell (#386)

* typo : Double Use of words (#374)

Fixing the typo double use of 'to', 'the'
Fixing typo of 'teh'

* Correcting typo (#341)

Typo in header, assuming it should be using instead of usubg

* Resolution for grammar issue #294 (#340)

* typo: teh -> the (#321)

* Update bot-builder-concept-luis.md (#380)

Fixed the missing links for speech and LUIS.

* Typo

dd => add

* Update image alignment in bot-builder-dotnet-sqk-quickstart.md

* Update bot-service-design-pattern-knowledge-base.md

The link was redirecting to the v3 bot builder templates which was missing the context of the paragraph.

* fix formatting and typo

* Resolve syncing conflicts from live-sync-work to live

* Update README.md

"About your contributions to Bot Framework content" link updated

* add additional fixes to this file

* Resolve syncing conflicts from live-sync-work to live (#399)

* Release 'live' 10-13-2018 (#1003)

* Release 'live' 10-15-2018 (#1011)

* Release 'live' 10-17-2018 (#1012)

* Release 'live' 10-18-2018 (#1016)

* Update bot-framework-rest-direct-line-3-0-api-reference.md (#431)

Add a comment about max activity size.

* Update bot-builder-enterprise-template-deployment.md (#427)

Added solution for common MSA app provisioning error.

* Update command to create metadata for qna maker (#405)

Previous command does not work. We should put one space.

* Wrongly used MD h2 title at "usage" section (#432)

Missing a space behind the second # at the "usage" section. The h2 style isn't rendered.
By adding a space behind ## it renders the title as an h2.

* Moved language-agnostic instruction (#460)

This snippet fell under a C# specific snippet, which made it unclear that it also applies to Node: 

"The `botFilePath` references the name of your bot, simply replace "yourbasicBot.bot" with your own bot name. To obtain the `botFileSecret` key, refer to [Bot File Encryption](https://aka.ms/bot-file-encryption) article on generating a key for your bot."

* Add "Dialog Lifecycle" header (#454)

The documentation opens with an overview and then flows right into discussing dialog lifecycle. I think adding a header would be a good to break that up and call more attention to the specific topic that is being discussed.

* Update issue templates

* Update issue templates

* Update issue templates

* Release 'live' 11-08-2018 (#1074)

* Replaced 'This topics' with 'This topic' (#512)

* Fix intro text for SDK v4 (#508)

As far as I know, the v4 of the C# and NodeJS SDK-s are not in preview anymore.