Update bot-builder-tutorial-basic-deploy.md
Even when the command and parameters name used are correct, the context given to the reader is poor, specially in the deployment section. You should explain first to the reader what are you going to do before start shooting commands. Also there are some instructions that are not clear such as:
"Select all the files and zip them up while in that folder, then run the command while still in that folder." What command?
The hierarchy of this section is not well structured. The title: Retrieve or create necessary IIS/Kudu files appears to be inside: Deploy via ARM template (with existing Resource Group). This confuses a lot, because it belong more to the part of deploying the code of the bot to an already prepared resource group.
I have proposed some additional context and changes in the hierarchy of the article.