Update bot-service-design-principles.md
Hi Kamran! I started on this first article - is this the right way to update for you?
Main bit around this piece, it only has the title named principles but inside of the article no principles are really shared, and parts of the content is outdated. 

Plus I've included the full guide, but not sure how to host it and if this is the way you've envisioned it. Since the alt text check is planned for next week I will have the final version of the Guide by then. 

It would be great to lift this bit "CUX" for the composer docs as well!
For the other articles following this one - I will update them less drastically (take out some legacy descriptions) and link to correct versions of Adaptive cards for example :)
I have planned to finish that next week for you!