Update docfx.json
There's another ms.date metadata existing here in the file under fileMetadata. It seems to be getting picked up by content metrics as the date instead of the globalMetadata ms.date. I need to check what the fileMetadata section is about, but in meantime, changing fileMetadata ms.date to the intended override date.